+++ src/TheNextWeek/bvh.h
@@ -31,9 +31,9 @@ class bvh_node : public hittable {
         const std::vector<shared_ptr<hittable>>& src_objects,
         size_t start, size_t end, double time0, double time1);
 
-    virtual bool hit(const ray& r, interval ray_t, hit_record& rec) const override;
+    bool hit(const ray& r, interval ray_t, hit_record& rec) const override;
 
-    virtual bool bounding_box(double time0, double time1, aabb& output_box) const override;
+    bool bounding_box(double time0, double time1, aabb& output_box) const override;
 
   public:
     shared_ptr<hittable> left;